+++ swift/Sources/LibSignalClient/Utils.swift
@@ -228,7 +228,8 @@ extension Optional where Wrapped: StringProtocol {
 }
 
 extension Array where Element == UInt8 {
-    internal func toHex() -> String {
+    /// Converts these bytes to (lowercase) hexadecimal.
+    public func toHex() -> String {
         var hex = [UInt8](repeating: 0, count: self.count * 2)
         hex.withUnsafeMutableBytes { hex in
             failOnError(
@@ -239,6 +240,26 @@ extension Array where Element == UInt8 {
                     self.count
                 )
             )
+        }
+        return String(decoding: hex, as: Unicode.UTF8.self)
+    }
+}
+
+extension Data {
+    /// Converts these bytes to (lowercase) hexadecimal.
+    public func toHex() -> String {
+        var hex = [UInt8](repeating: 0, count: self.count * 2)
+        hex.withUnsafeMutableBytes { hex in
+            self.withUnsafeBytes { input in
+                failOnError(
+                    signal_hex_encode(
+                        hex.baseAddress?.assumingMemoryBound(to: CChar.self),
+                        hex.count,
+                        input.baseAddress?.assumingMemoryBound(to: UInt8.self),
+                        input.count
+                    )
+                )
+            }
         }
         return String(decoding: hex, as: Unicode.UTF8.self)
     }